Add function to shuffle list of names
Add function shuffle_list in file meeting_part_manager.py. It randomizes the list of names in a way that the same name never repeats in a column.
Altered code so that 2 lists of names are stored (friday and tuesday).
Also added an index mapping each entry in the lists to entries from the complete list of MeetingParts.