Modify create_parts function to handle edge cases
Slightly alter code in the create_parts() function. This will account for columns in the input that are incomplete and ignore those columns, by checking the MeetingPart.dead variable. Should lead to a much more robust program that will handle edge cases elegantly and simply ignore unusable data instead of printing an error message to the screen